core/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.2
fastapi==0.104.1
h11==0.14.0
idna==3.4
Markdown==3.4.3
outcome==1.2.0
praw==7.7.0
prawcore==2.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
APScheduler==3.10.4
alembic==1.12.1
SQLAlchemy==2.0.23
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
update-checker==0.18.0
urllib3==2.0.3
uvicorn[standard]==0.24.0
wcwidth==0.2.6
websocket-client==1.6.0
wsproto==1.2.0
python-dotenv==1.0.0
